Model/add categories

+++ src/main/java/com/cogent/ecommerce/controller/CategoryController.java
@@ -1,6 +1,7 @@
 package com.cogent.ecommerce.controller;
 
 import com.cogent.ecommerce.model.Category;
+import com.cogent.ecommerce.security.AuthService;
 import com.cogent.ecommerce.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -13,16 +14,28 @@ import java.util.List;
 @CrossOrigin(origins = "http://localhost:4200")
 public class CategoryController {
 
-    //TODO: FOR SECURITY THESE NEED TO ONLY BE ACCESSIBLE TO ADMINS
+    //TODO: FOR SECURITY THESE NEED TO ONLY BE ACCESSIBLE TO ADMINS--DONE
     @Autowired
     CategoryService categoryService;
+
+    @Autowired
+    AuthService authService;
 
     @GetMapping("/getCategory")
     public List<Category> getCategory() {
         return categoryService.getAllCategories();
     }
     @PostMapping("/addCategory")
-    public ResponseEntity<?> addCategory(@RequestBody Category category) {
+    public ResponseEntity<?> addCategory(@RequestHeader("Authorization") String authHeader,@RequestBody Category category) {
+
+        String token = authHeader.substring(7);
+        if(!authService.checkIfAdmin(token)){
+            return ResponseEntity.badRequest().body("Not Authorized");
+        }
+
+        if(categoryService.getCategoryByName(category.getName()).isPresent()){
+            return ResponseEntity.badRequest().body("Category Exists");
+        }
         if(categoryService.getCategoryById(category.getId()).isPresent()){
             return ResponseEntity.badRequest().body("Category Exists");
         }
@@ -32,11 +45,22 @@ public class CategoryController {
 
     //will not work until bidirectional many to many added-->next task so we can remove categories and it will remove product ref to that cat;
     @DeleteMapping("/deleteCategory/{id}")
-    public ResponseEntity<?> deleteCategory(@PathVariable int id){
-
+    public ResponseEntity<?> deleteCategory(@RequestHeader("Authorization") String authHeader,@PathVariable int id){
+        String token = authHeader.substring(7);
+        if(!authService.checkIfAdmin(token)){
+            return ResponseEntity.badRequest().body("Not Authorized");
+        }
         return ResponseEntity.ok().body(categoryService.deleteCategory(id));
 
     }
 
+    @PutMapping("/updateCategory/{id}/{name}")
+    public ResponseEntity<?> updateCategory(@RequestHeader("Authorization") String authHeader,@PathVariable int id, @PathVariable String name){
+        String token = authHeader.substring(7);
+        if(!authService.checkIfAdmin(token)){
+            return ResponseEntity.badRequest().body("Not Authorized");
+        }
+        return ResponseEntity.ok().body(categoryService.updateCategory(id, name));
+    }
 
 }
